added base style variables
