+++ src/client/src/components/ProductCardGrid/styles.module.scss
@@ -1,10 +1,11 @@
+@import 'styles/_variables.scss';
+
 .cardTitle {
-  font-family: 'Poppins', sans-serif !important;
   font-weight: 800 !important;
   line-height: 24px;
   height: 50px;
   letter-spacing: 0.5px;
-  color: #223263;
+  color: $text-heading-color;
   overflow: hidden;
   text-overflow: ellipsis;
   text-align: center;
@@ -22,12 +23,11 @@
 }
 
 .cardBody {
-  background-color: #f6f6f6;
+  background-color: $background-color-neutral;
 }
 
 .cardPricing {
-  font-family: 'Poppins', sans-serif !important;
-  color: #40bfff;
+  color: $primary-color;
   font-weight: 800;
   letter-spacing: 0.5px;
   font-size: 18px;
