+++ src/client/src/index.tsx
@@ -3,6 +3,7 @@ import ReactDOM from 'react-dom';
 
 import App from 'containers/App';
 
+import 'styles/_variables.scss';
 import 'styles/common.scss';
 
 ReactDOM.render(<App />, document.getElementById('root'));
