--- src/client/src/pages/Home/containers/HomeProductListings/styles.module.scss
@@ -1,17 +1,0 @@
-.categoryGrid {
-  display: grid;
-  grid-template-columns: repeat(1, minmax(150px, 1fr));
-  grid-gap: 20px;
-
-  @media (min-width: 521px) {
-    grid-template-columns: repeat(2, minmax(180px, 1fr));
-  }
-
-  @media (min-width: 768px) {
-    grid-template-columns: repeat(3, minmax(200px, 1fr));
-  }
-
-  @media (min-width: 1280px) {
-    grid-template-columns: repeat(4, minmax(200px, 1fr));
-  }
-}
