+++ src/client/src/pages/Home/index.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { Container } from 'react-bootstrap';
 
 import HomeBanner from 'pages/Home/components/HomeBanner';
-import HomeProductListings from 'pages/Home/containers/HomeProductListings';
+import HomeProductListings from 'pages/Home/components/HomeProductListings';
 import ListingsFilters from 'components/ListingsFilters';
 import { ProductListingsProvider, ProductListingsConsumer } from 'contexts/ProductListingsContext';
 
@@ -10,6 +10,7 @@ const Home: React.FC = (): React.ReactElement => {
   return (
     <Container className="mt-4 mb-4">
       <HomeBanner />
+
       <ProductListingsProvider>
         <ProductListingsConsumer>
           {({ data, filteredData, updateFilter }) => (
