+++ src/client/src/styles/_variables.scss
@@ -1,0 +1,7 @@
+$primary-color: #40bfff;
+$secondary-color: #bcddfe;
+$text-color: #373737;
+$text-heading-color: #223263;
+$text-label-color: #22262a;
+$accent-color: #ff4858;
+$background-color-neutral: #f6f6f6;
